requirements.txt
# LangGraph core
langgraph>=0.2.0
langgraph-prebuilt>=0.1.0
langgraph-checkpoint-sqlite>=1.0.0
langchain-core>=0.3.0

# LLM provider integrations (LangChain)
langchain-openai>=0.2.0
langchain-anthropic>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0

# Provider SDKs (direct)
openai>=1.50.0
anthropic>=0.39.0
google-generativeai>=0.8.0
huggingface-hub>=0.25.0

# Observability
langsmith>=0.1.0

# API + UI
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
gradio>=4.40.0
sse-starlette>=2.0.0

# Utilities
httpx>=0.27.0
python-dotenv>=1.0.0
